# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Synconomics, build with go",
        "title": "Synconomics API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "mengambil semua data transaksi yang ada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "ambil semua daftar transaksi",
                "responses": {
                    "200": {
                        "description": "Sukses mengambil data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.TransactionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan data transaksi baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Membuat transaksi baru",
                "parameters": [
                    {
                        "description": "Transaction Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil informasi lengkap terkait satu transaksi menggunakan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Mendapatkan detail transaksi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah kolom dari sebuah entitas transaksi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Memperbarui data transaksi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menghapus transaksi dari database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Menghapus transaksi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "business_id",
                "payment_method",
                "status",
                "total_amount"
            ],
            "properties": {
                "business_id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "completed",
                        "cancelled"
                    ]
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "dto.TransactionResponse": {
            "type": "object",
            "properties": {
                "business_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "number"
                },
                "transaction_date": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateTransactionRequest": {
            "type": "object",
            "properties": {
                "payment_method": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "completed",
                        "cancelled"
                    ]
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "models.AIMessage": {
            "type": "object",
            "properties": {
                "aisession": {
                    "$ref": "#/definitions/models.AISession"
                },
                "aisessionID": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.AIResult": {
            "type": "object",
            "properties": {
                "aisession": {
                    "$ref": "#/definitions/models.AISession"
                },
                "aisessionID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "result_type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.AISession": {
            "type": "object",
            "properties": {
                "aimessages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AIMessage"
                    }
                },
                "airesults": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AIResult"
                    }
                },
                "business": {
                    "$ref": "#/definitions/models.Business"
                },
                "businessID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.AISessionType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "models.Business": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "aisessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AISession"
                    }
                },
                "businessMetrics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BusinessMetric"
                    }
                },
                "businessScores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BusinessScore"
                    }
                },
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "expenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Expense"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "instagram": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "logo_url": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Product"
                    }
                },
                "supplyOffers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SupplyOffer"
                    }
                },
                "supplyRequests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SupplyRequest"
                    }
                },
                "tiktok": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "description": "Foreign Key to User",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "userID": {
                    "type": "integer"
                },
                "website": {
                    "type": "string"
                },
                "whatsapp": {
                    "type": "string"
                }
            }
        },
        "models.BusinessMetric": {
            "type": "object",
            "properties": {
                "business": {
                    "$ref": "#/definitions/models.Business"
                },
                "business_id": {
                    "type": "integer"
                },
                "calculated_at": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "net_profit": {
                    "type": "number"
                },
                "sales_growth": {
                    "type": "number"
                },
                "stock_turnover": {
                    "type": "number"
                },
                "total_expense": {
                    "type": "number"
                },
                "total_revenue": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.BusinessScore": {
            "type": "object",
            "properties": {
                "business_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "insight": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Expense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "business": {
                    "$ref": "#/definitions/models.Business"
                },
                "businessID": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Product": {
            "type": "object",
            "properties": {
                "business": {
                    "$ref": "#/definitions/models.Business"
                },
                "business_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "min_stock": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "supplyOffers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SupplyOffer"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.SupplyOffer": {
            "type": "object",
            "properties": {
                "business": {
                    "$ref": "#/definitions/models.Business"
                },
                "business_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "supplyMatches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SupplyMatch"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SupplyRequest": {
            "type": "object",
            "properties": {
                "business": {
                    "$ref": "#/definitions/models.Business"
                },
                "businessID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.SupplyRequestStatus"
                },
                "supplyMatches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SupplyMatch"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Transaction": {
            "type": "object",
            "properties": {
                "business": {
                    "$ref": "#/definitions/models.Business"
                },
                "business_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.TransactionStatus"
                },
                "total_amount": {
                    "type": "number"
                },
                "transactionItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionItem"
                    }
                },
                "transaction_date": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.TransactionItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "productID": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "transaction": {
                    "$ref": "#/definitions/models.Transaction"
                },
                "transactionID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.TransactionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "completed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusCompleted",
                "StatusCancelled"
            ]
        },
        "models.User": {
            "type": "object",
            "properties": {
                "aisessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AISession"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "business": {
                    "$ref": "#/definitions/models.Business"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "google_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "productSearchLogs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductSearchLog"
                    }
                },
                "provider": {
                    "type": "string"
                },
                "replies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Reply"
                    }
                },
                "threads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Thread"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JSON Web Token. Example: \"Bearer eyJhb...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}